# schemas/generated/orchestration-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tasker Web API",
    "description": "REST API for the Tasker workflow orchestration system",
    "contact": {
      "name": "Tasker Systems",
      "email": "support@tasker-systems.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "API version 1"
    }
  ],
  "paths": {
...
    "/v1/tasks": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "List tasks with pagination and filtering: GET /v1/tasks",
        "description": "Returns a paginated list of tasks with execution context using batch SQL functions.\nUsed by workers and administrators for task monitoring and management.\n\n**Required Permission:** `tasks:list`",
        "operationId": "list_tasks",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number (default: 1)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "description": "Items per page (default: 25, max: 100)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "namespace",
            "in": "query",
            "description": "Filter by namespace",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter by status",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "initiator",
            "in": "query",
            "description": "Filter by initiator",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "source_system",
            "in": "query",
            "description": "Filter by source system",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of tasks",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "api_key_auth": []
          }
        ],
        "x-required-permission": "tasks:list"
      },
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Create a new task: POST /v1/tasks",
        "description": "This is the critical endpoint for TAS-40 worker integration.\nWorkers call this endpoint to create tasks and receive UUIDs for tracking.\n\nReturns the same `TaskResponse` shape as GET /v1/tasks/{uuid}, following REST best\npractices where create operations return the same representation as read operations.\n\n**Required Permission:** `tasks:create`",
        "operationId": "create_task",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Task created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "api_key_auth": []
          }
        ],
        "x-required-permission": "tasks:create"
      }
    },
    "/v1/tasks/{uuid}": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Get task details: GET /v1/tasks/{uuid}",
        "description": "Returns comprehensive task information including execution context and step readiness.\nUsed by workers for task correlation and status tracking.\n\n**Required Permission:** `tasks:read`",
        "operationId": "get_task",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "description": "Task UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Task not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "api_key_auth": []
          }
        ],
        "x-required-permission": "tasks:read"
      },
      "delete": {
        "tags": [
          "tasks"
        ],
        "summary": "Cancel a task: DELETE /v1/tasks/{uuid}",
        "description": "Cancels a task if it's in a cancellable state and triggers orchestration events.\nReturns the updated task with execution context and step information.\n\n**Required Permission:** `tasks:cancel`",
        "operationId": "cancel_task",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "description": "Task UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task cancelled successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "400": {
            "description": "Task cannot be cancelled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Task not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "api_key_auth": []
          }
        ],
        "x-required-permission": "tasks:cancel"
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "NotFound"
            ],
            "properties": {
              "NotFound": {
                "type": "object",
                "required": [
                  "message"
                ],
                "properties": {
                  "message": {
                    "type": "string"
                  }
                }
              }
            }
          },
          {
            "type": "string",
            "enum": [
              "Forbidden"
            ]
          },
          {
            "type": "string",
            "enum": [
              "Unauthorized"
            ]
          },
          {
            "type": "object",
            "required": [
              "BadRequest"
            ],
            "properties": {
              "BadRequest": {
                "type": "object",
                "required": [
                  "message"
                ],
                "properties": {
                  "message": {
                    "type": "string"
                  }
                }
              }
            }
          },
          {
            "type": "object",
            "description": "TAS-154: Conflict - resource already exists (e.g., duplicate identity hash)",
            "required": [
              "Conflict"
            ],
            "properties": {
              "Conflict": {
                "type": "object",
                "description": "TAS-154: Conflict - resource already exists (e.g., duplicate identity hash)",
                "required": [
                  "message"
                ],
                "properties": {
                  "message": {
                    "type": "string"
                  }
                }
              }
            }
          },
          {
            "type": "string",
            "enum": [
              "ServiceUnavailable"
            ]
          },
          {
            "type": "string",
            "enum": [
              "Timeout"
            ]
          },
          {
            "type": "string",
            "enum": [
              "CircuitBreakerOpen"
            ]
          },
          {
            "type": "object",
            "description": "TAS-75: System backpressure active - includes Retry-After header",
            "required": [
              "Backpressure"
            ],
            "properties": {
              "Backpressure": {
                "type": "object",
                "description": "TAS-75: System backpressure active - includes Retry-After header",
                "required": [
                  "reason",
                  "retry_after_seconds"
                ],
                "properties": {
                  "reason": {
                    "type": "string"
                  },
                  "retry_after_seconds": {
                    "type": "integer",
                    "format": "int64",
                    "minimum": 0
                  }
                }
              }
            }
          },
          {
            "type": "object",
            "required": [
              "DatabaseError"
            ],
            "properties": {
              "DatabaseError": {
                "type": "object",
                "required": [
                  "operation"
                ],
                "properties": {
                  "operation": {
                    "type": "string"
                  }
                }
              }
            }
          },
          {
            "type": "object",
            "required": [
              "AuthenticationError"
            ],
            "properties": {
              "AuthenticationError": {
                "type": "object",
                "required": [
                  "reason"
                ],
                "properties": {
                  "failure_context": {
                    "oneOf": [
                      {
                        "type": "null"
                      },
                      {
                        "$ref": "#/components/schemas/AuthFailureContext",
                        "description": "Machine-readable failure context for gateway signaling headers"
                      }
                    ]
                  },
                  "reason": {
                    "type": "string"
                  }
                }
              }
            }
          },
          {
            "type": "object",
            "required": [
              "AuthorizationError"
            ],
            "properties": {
              "AuthorizationError": {
                "type": "object",
                "required": [
                  "reason"
                ],
                "properties": {
                  "failure_context": {
                    "oneOf": [
                      {
                        "type": "null"
                      },
                      {
                        "$ref": "#/components/schemas/AuthFailureContext",
                        "description": "Machine-readable failure context for gateway signaling headers"
                      }
                    ]
                  },
                  "reason": {
                    "type": "string"
                  }
                }
              }
            }
          },
          {
            "type": "object",
            "required": [
              "InvalidUuid"
            ],
            "properties": {
              "InvalidUuid": {
                "type": "object",
                "required": [
                  "uuid"
                ],
                "properties": {
                  "uuid": {
                    "type": "string"
                  }
                }
              }
            }
          },
          {
            "type": "string",
            "enum": [
              "JsonError"
            ]
          },
          {
            "type": "string",
            "enum": [
              "Internal"
            ]
          }
        ],
        "description": "Web API specific errors with HTTP status code mappings"
      },
...
      "PaginationInfo": {
        "type": "object",
        "description": "Pagination metadata",
        "required": [
          "page",
          "per_page",
          "total_count",
          "total_pages",
          "has_next",
          "has_previous"
        ],
        "properties": {
          "has_next": {
            "type": "boolean"
          },
          "has_previous": {
            "type": "boolean"
          },
          "page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "per_page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "total_count": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "total_pages": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "TaskListResponse": {
        "type": "object",
        "description": "Task list response with pagination",
        "required": [
          "tasks",
          "pagination"
        ],
        "properties": {
          "pagination": {
            "$ref": "#/components/schemas/PaginationInfo"
          },
          "tasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskResponse"
            }
          }
        }
      },
      "TaskRequest": {
        "type": "object",
        "description": "TaskRequest represents an incoming request to create and execute a task\nThis is the primary routing input that identifies which handler should process the task\nand contains all the information needed to create a Task instance from a NamedTask template",
        "required": [
          "name",
          "namespace",
          "version",
          "context",
          "initiator",
          "source_system",
          "reason",
          "tags",
          "requested_at",
          "correlation_id"
        ],
        "properties": {
          "context": {
            "description": "Context data required for task execution (Rails: context)"
          },
          "correlation_id": {
            "type": "string",
            "format": "uuid",
            "description": "Correlation ID for distributed tracing (auto-generated if not provided)\nTAS-29: Enables end-to-end request tracking across orchestration and workers"
          },
          "idempotency_key": {
            "type": [
              "string",
              "null"
            ],
            "description": "TAS-154: Optional caller-provided idempotency key.\n\nWhen provided, this key is used to compute the task's identity_hash,\noverriding the named task's default identity strategy.\n\nUse cases:\n- Override STRICT strategy with a custom key\n- Provide required key for CALLER_PROVIDED strategy\n- Time-bucketed keys for windowed deduplication (e.g., `\"job-2026-01-20-09\"`)\n\nThe key is combined with the named_task_uuid to prevent collisions across\ndifferent named tasks."
          },
          "initiator": {
            "type": "string",
            "description": "The entity or system that initiated this task request (Rails: initiator)"
          },
          "name": {
            "type": "string",
            "description": "The name of the task to be performed (Rails: name)"
          },
          "namespace": {
            "type": "string",
            "description": "The namespace of the task to be performed (Rails: namespace, default: \"default\")"
          },
          "options": {
            "type": [
              "object",
              "null"
            ],
            "description": "Custom options that override task template defaults (Rust extension)",
            "additionalProperties": {},
            "propertyNames": {
              "type": "string"
            }
          },
          "parent_correlation_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Optional parent correlation ID for nested/chained workflow relationships\nTAS-29: Enables tracking of workflow hierarchies"
          },
          "priority": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Priority for task execution (higher values = higher priority). Default: 0"
          },
          "reason": {
            "type": "string",
            "description": "The reason why this task was requested (Rails: reason)"
          },
          "requested_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the task was initially requested (Rails: requested_at)"
          },
          "source_system": {
            "type": "string",
            "description": "The system from which this task originated (Rails: source_system)"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Tags associated with this task for categorization or filtering (Rails: tags)"
          },
          "version": {
            "type": "string",
            "description": "The version of the task to be performed (Rails: version, default: \"0.1.0\")"
          }
        }
      },
      "TaskResponse": {
        "type": "object",
        "description": "Task details response with execution context and step readiness information",
        "required": [
          "task_uuid",
          "name",
          "namespace",
          "version",
          "status",
          "created_at",
          "updated_at",
          "context",
          "initiator",
          "source_system",
          "reason",
          "correlation_id",
          "total_steps",
          "pending_steps",
          "in_progress_steps",
          "completed_steps",
          "failed_steps",
          "ready_steps",
          "execution_status",
          "recommended_action",
          "completion_percentage",
          "health_status",
          "steps"
        ],
        "properties": {
          "completed_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "completed_steps": {
            "type": "integer",
            "format": "int64"
          },
          "completion_percentage": {
            "type": "number",
            "format": "double"
          },
          "context": {},
          "correlation_id": {
            "type": "string",
            "format": "uuid"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "execution_status": {
            "type": "string"
          },
          "failed_steps": {
            "type": "integer",
            "format": "int64"
          },
          "health_status": {
            "type": "string"
          },
          "in_progress_steps": {
            "type": "integer",
            "format": "int64"
          },
          "initiator": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "namespace": {
            "type": "string"
          },
          "parent_correlation_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "pending_steps": {
            "type": "integer",
            "format": "int64"
          },
          "priority": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "ready_steps": {
            "type": "integer",
            "format": "int64"
          },
          "reason": {
            "type": "string"
          },
          "recommended_action": {
            "type": "string"
          },
          "source_system": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "steps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StepReadinessStatus"
            }
          },
          "tags": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "task_uuid": {
            "type": "string"
          },
          "total_steps": {
            "type": "integer",
            "format": "int64"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "version": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_key_auth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}